move color/term functions to bokeh.util.terminal
