+++ tests/examples/test_examples.py
@@ -10,12 +10,12 @@ import signal
 from os.path import dirname, exists, split
 
 from bokeh.server.callbacks import NextTickCallback, PeriodicCallback, TimeoutCallback
-from tests.plugins.utils import trace, info, fail, ok, red, warn, white
 from tests.plugins.screenshot import get_screenshot
 from tests.plugins.image_diff import image_diff
 
 from bokeh.client import push_session
 from bokeh.command.util import build_single_handler_application
+from bokeh.util.terminal import trace, info, fail, ok, red, warn, white
 
 @pytest.mark.examples
 def test_js_examples(js_example, example, report):
