+++ tests/plugins/screenshot.py
@@ -6,7 +6,7 @@ import json
 import subprocess
 
 from os.path import abspath, dirname, join, pardir, split
-from .utils import trace, fail
+from bokeh.util.terminal import trace, fail
 
 TOP_PATH = abspath(join(split(__file__)[0], pardir, pardir))
